Layout: Remove hover highlight
The highlight wasn't really useful and can be problematic when using
touch screens as the mouse will not move off of a cell and therefore
override the preview/program border color.

Technically it could be changed so that the custom border color of the
cell takes priority over its hover status but for now we'll just leave
it at this.

Closes #16