Durchblick: Cache last saved layout
When durchblick gets hidden it deletes the layout to release source
references. This also means that we don't have a layout when saving the
next time and the window is still invisible as the layout is only loaded
from disk when showing the window again. To prevent the loss of layouts
when saving hidden durchblick windows we just cache the last layout and
save that when the window is invisible.